+++ example/lib/main.dart
@@ -11,6 +11,7 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'Flutter Example',
       navigatorKey: Routerino.navigatorKey,
+      navigatorObservers: [RouterinoObserver()],
       home: RouterinoHome(
         builder: () => HomePage(),
       ),
@@ -39,6 +40,12 @@ class HomePage extends StatelessWidget {
                 print('RESULT: $result (${result.runtimeType})');
               },
               child: Text('Pick a number'),
+            ),
+            ElevatedButton(
+              onPressed: () {
+                context.push(() => RemoveBeneathTest());
+              },
+              child: Text('Remove beneath'),
             ),
             ElevatedButton(
               onPressed: () {
@@ -82,7 +89,7 @@ class LoginPage extends StatelessWidget {
               ),
               ElevatedButton(
                 onPressed: () {
-                  context.push(() => RegisterPage(), transition: RouterinoTransition.fade);
+                  context.push(() => RegisterPage(), transition: RouterinoTransition.fade());
                 },
                 child: Text('Register'),
               ),
@@ -142,3 +149,50 @@ class PickNumberPage extends StatelessWidget with PopsWithResult<int> {
   }
 }
 
+class RemoveBeneathTest extends StatelessWidget {
+  const RemoveBeneathTest({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Column(
+        children: [
+          const Text('Remove Beneath: A'),
+          ElevatedButton(
+            onPressed: () {
+              context.push(() => RemoveBeneathTest2());
+            },
+            child: Text('Next'),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class RemoveBeneathTest2 extends StatelessWidget {
+  const RemoveBeneathTest2({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Column(
+        children: [
+          const Text('Remove Beneath: B'),
+          ElevatedButton(
+            onPressed: () {
+              context.removeRoute(RemoveBeneathTest);
+            },
+            child: Text('Remove A'),
+          ),
+          ElevatedButton(
+            onPressed: () {
+              context.pop();
+            },
+            child: Text('Pop'),
+          ),
+        ],
+      ),
+    );
+  }
+}